Moved city skeleton generation into CityLevelUtils.
Can be used by both city generation and wilderness city revision.
