+++ OpenTESArena/src/World/ExteriorLevelData.cpp
@@ -40,30 +40,15 @@ ExteriorLevelData ExteriorLevelData::loadCity(const LocationDefinition &location
 	int currentDay, int starCount, const std::string &infName, SNInt gridWidth, WEInt gridDepth,
 	const MiscAssets &miscAssets, TextureManager &textureManager)
 {
-	const BufferView2D<const MIFFile::VoxelID> levelFLOR = level.getFLOR();
-	const BufferView2D<const MIFFile::VoxelID> levelMAP1 = level.getMAP1();
-	const BufferView2D<const MIFFile::VoxelID> levelMAP2 = level.getMAP2();
-	const WEInt levelWidth = levelFLOR.getWidth();
-	const SNInt levelDepth = levelFLOR.getHeight();
-
 	// Create temp voxel data buffers and write the city skeleton data to them. Each city
 	// block will be written to them as well.
-	Buffer2D<uint16_t> tempFlor(levelWidth, levelDepth);
-	Buffer2D<uint16_t> tempMap1(levelWidth, levelDepth);
-	Buffer2D<uint16_t> tempMap2(levelWidth, levelDepth);
-
-	for (WEInt x = 0; x < levelWidth; x++)
-	{
-		for (SNInt z = 0; z < levelDepth; z++)
-		{
-			const MIFFile::VoxelID srcFlorVoxel = levelFLOR.get(x, z);
-			const MIFFile::VoxelID srcMap1Voxel = levelMAP1.get(x, z);
-			const MIFFile::VoxelID srcMap2Voxel = levelMAP2.get(x, z);
-			tempFlor.set(x, z, srcFlorVoxel);
-			tempMap1.set(x, z, srcMap1Voxel);
-			tempMap2.set(x, z, srcMap2Voxel);
-		}
-	}
+	Buffer2D<uint16_t> tempFlor(gridDepth, gridWidth);
+	Buffer2D<uint16_t> tempMap1(gridDepth, gridWidth);
+	Buffer2D<uint16_t> tempMap2(gridDepth, gridWidth);
+	CityLevelUtils::writeSkeleton(level,
+		BufferView2D(tempFlor.get(), tempFlor.getWidth(), tempFlor.getHeight()),
+		BufferView2D(tempMap1.get(), tempMap1.getWidth(), tempMap1.getHeight()),
+		BufferView2D(tempMap2.get(), tempMap2.getWidth(), tempMap2.getHeight()));
 
 	// Get the city's seed for random chunk generation. It is modified later during
 	// building name generation.
